refactor: rename sdkClient.mjs to setupDashClient.mjs
